pentagon1024: small build fix

+++ Kernel/platform-pentagon1024/main.c
@@ -34,7 +34,7 @@ void plt_interrupt(void)
  *	the runtime
  */
 
-int strlen(const char *p)
+size_t strlen(const char *p)
 {
   int len = 0;
   while(*p++)
